+++ Lab3/Task3/cfd/jacobi.cpp
@@ -3,7 +3,7 @@
 #include "jacobi.h"
 
 
-void jacobistep(double *psinew, double *psi, int m, int n)
+void jacobistep(float *psinew, float *psi, int m, int n)
 {
 	int i, j;
   
@@ -15,12 +15,12 @@ void jacobistep(double *psinew, double *psi, int m, int n)
 }
 
 
-double deltasq(double *newarr, double *oldarr, int m, int n)
+float deltasq(float *newarr, float *oldarr, int m, int n)
 {
 	int i, j;
 
-	double dsq=0.0;
-	double tmp;
+	float dsq=0.0;
+	float tmp;
 
 	for(i=1;i<=m;i++)
 	{
