+++ Lab3/Task3/cfd_cl/boundary.cpp
@@ -3,7 +3,7 @@
 
 //grid is parallelised in the x direction
 
-void boundarypsi(double *psi, int m, int n, int b, int h, int w)
+void boundarypsi(float *psi, int m, int n, int b, int h, int w)
 {
 	int i,j;
