+++ Lab3/Task3/cfd_cl/boundary.h
@@ -1,1 +1,1 @@
-void boundarypsi(double *psi, int m, int n, int b, int h, int w);
+void boundarypsi(float *psi, int m, int n, int b, int h, int w);
